Micrositio/requirements.txt
asgiref==3.2.3
astroid==2.3.3
colorama==0.4.3
Django==2.1.15
django-cors-headers==3.2.1
django-pyodbc==1.1.3
django-pyodbc-azure==2.1.0.0
djangorestframework==3.11.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pylint==2.4.4
pyodbc==4.0.30
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
whitenoise==5.0.1
wrapt==1.11.2